@@ -20,7 +20,7 @@
 		"linux"
 	],
 	"engines": {
-		"node": ">=16.5"
+		"node": ">=16.6"
 	},
 	"maintainers": [
 		{
@@ -77,13 +77,13 @@
 		"@types/browserify": "^12.0.37",
 		"@types/express": "^4.17.13",
 		"@types/fs-extra": "^9.0.12",
-		"@types/node": "^16.4.7",
+		"@types/node": "^16.4.10",
 		"@types/nodemailer": "^6.4.4",
 		"@types/nodemailer-smtp-transport": "^2.7.5",
 		"@types/sprintf-js": "^1.1.2",
 		"@types/ws": "^7.4.7",
 		"browserify": "^17.0.0",
-		"eslint": "^7.31.0",
+		"eslint": "^7.32.0",
 		"typescript": "^4.3.5",
 		"watchify": "^4.0.0"
 	},
