@@ -4,14 +4,14 @@
 \*****************************************************************************/
 
 import $ = require('../runtime')
-import Battle = require('../battle')
 import db = require('../db')
 import { Armor, Coin, Magic, Poison, Weapon } from '../items'
 import { bracket, cat, death, display, log, vt, weapon } from '../lib'
-import { dungeon, elemental } from '../npc'
 import { PC } from '../pc'
+import { dungeon, elemental } from '../npc'
 import { checkXP, input } from '../player'
 import { cuss, dice, int, money, sprintf, titlecase, whole } from '../sys'
+import Battle = require('./battle')
 
 module Party {
 
